Run the ceph osd execs with the shell provider
Some of the exec resources that are used in the ceph::osd class require
the use of the shell provider, since they use some builtin functions
like 'if'.

It's not the most elegant solution and it would be better to rewrite the
whole thing as a provider, once we have upgraded to puppet 7.

Bug: T330151
Change-Id: I868b074bc62b581137ff568217bf616ed630a402