deployment: Support making k8s deploys db section aware
Why:
In T340843 and linked from it tasks, we identified that the current way
of hardcoding dbproxy hosts isn't sustainable long term and settled on
an approach of allowing egress functionality from k8s deployed
applications based on TCP ports.

In Idf9d64a026ed33c170066d7c84af8c9d366d7ada I 've added the
functionality to the base sextant modules, but in order to not duplicate
the mariadb section stanzas in charts, we need a glue point. That point
is currently best done  in the same spirit as kafka_brokers and
services proxies

What:
Add a deployment stanza to populate a mariadb.section_ports key that has
all the section_ports as defined in profile::mariadb::section_ports

Hosts: auto
Bug: T340843
Signed-off-by: Alexandros Kosiaris <akosiaris@wikimedia.org>
Change-Id: Id6172c7cbd8e13dbe6571ea0d91c6c80ba89b0ce