Disable the ceph fsid mismatch check
This safety check was not working as expected and was blocking the
subsequent prepare and activation steps. Disabling the check for now,
but we will rewrite it when we refactor this class as a custom provider.

The safety check does not really offer any value to us at this point
anyway, since we're only applying this defined type to a single role and
there is only one cluster fsid that we're considering.

Bug: T330151
Change-Id: I225dee25a44eb068674123e2a2d88dcfd2ae01c8