Fix the wwn of sas drives used for the cephosd servers
We now use the wwn of a drive when configuring Ceph OSDs, but for SAS
drives this can be a little complicated. The WWN reported by the HBA is
the baseline address for the device. This is always divisible by four.
The next two addresses above this represent the two possible ports for
connection to it. The third byte above represents the LUN contained in
the hard drive.

It is this LUN address that is detected by Linux and appears in
/dev/disk/by-id/wwn-* and therefore we need to add three bytes to the
wwn that comes from perccli64 before we send it to the ceph-volume lv
prepare command.

Bug: T330151
Change-Id: I29fd5ca8c5aa4572a089cf485e6b93073e88b753