tools-static: Hide more Cloudflare response headers
CF-Cdnjs-Via is presumably just as uninteresting as CF-Cache-Status and
CF-Ray. NEL and Report-To point to a Cloudflare NEL service; we
definitely don’t want our users to report connectivity problems to
Cloudflare, both because that doesn’t make sense and because it reveals
their IP. Alt-Svc seems to indicate that Cloudflare speaks HTTP3 and the
browser could upgrade to it; I don’t know if that’s true in our
infrastructure (no version of curl I have access to supports --http3),
but if it is, it should be indicated via a header we add ourselves, not
via one we just copy from Cloudflare.

(I’ve normalized the header names to Title-Case, with the acronyms CF
and NEL in all-caps, to match the existing ones; Cloudflare seems to
respond with lowercase header names, but I assume nginx matches them
case-insensitively.)

Change-Id: I87d78de57280bc64fbdc6b76b28f985290b0ee5b