varnish: handle varnish-frontend-hospital crash with a (null) line
When varnish-frontend-hospital gets a log line with (null) instead of
the vcl_id and origin_server, it results in a crash:

    0 Backend_health - (null) Went sick -------H 2 3 5 0.000000 0.000000
        ValueError: not enough values to unpack (expected 2, got 1)

It crashes because it gets a ['(null')] and can't assign vcl_id and
origin_server on the split.

When this happens, we should gracefully return instead of letting
varnish-hospital crash and restart.

Bug: T342566
Change-Id: Ieb8ad3d46e3954e53fa0f556124b7b91cc10a0b2