Exclude nagios checks of tmpfs mounts on cephosd servers
The cephosd servers mount a tmpfs file system under /var/lib/ceph for
each OSD daemon. The nagios user doesn't have the rights to read this
filesystem, so without this exclusion we would have alerts.

We will be able to measure the amount of memeory available to the host
via other means, so the best solution is to exclude the tmpfs type from
being monitored on this host.

Bug: T330151
Change-Id: I0ff9c34b18ed121c6f54fa8d359839407d376e7d