role::kafka::main: increase worker threads for kafka-main1001
We are now using only 4 threads (num.io.threads) to depool msgs
from the kafka input queue, process the data that they carry
and store them on page-cache/disk. The original idea was to
follow the same number of disks present on the node, but
I believe that this is causing backlogs in the brokers
that don't have enough capacity to process incoming msgs.

Apply the same value, 8, to the three settings that we configure
in heira (based on the above assumptions) on kafka-main1001
to measure their impact in performance.

Bug: T341558
Change-Id: I134ab2abc304d0ae2ce789c76d2419d4df84f000