profile::kafka::broker: fix settings passed to the confluent class
The num_io_threads and num_recovery_threads_per_data_dir are hiera
values that we configure for each kafka cluster, but we don't pass
it to the confluent class so they run with their default.

For example, num_io_threads represents the amount of threads working
on kafka msgs and persisting them on disk, and it is currently set
to 1 on all clusters.

This change sets the correct behavior in the profile, and comments
all the affected hiera settings to avoid a widespread roll out.

Only kafka-main1001 is updated, to see if the changes will have
any positive effect on it.

Bug: T341558
Change-Id: I7bb48d1acdd818d795ec07a9296d911596f2d862