profile::logstash: allow more Istio ingress gateway logs
In T313099 we applied a heavy drop filter to the output
of the envoy-proxy's containers, to avoid spammy messages
on logstash. The ML and ServiceOps teams now run various
kind of *-proxy, like istio-proxy, that would benefit
from having a complete access log stored in logstash/Elastic.

The Istio Gateway pods run a container called istio-proxy,
that is the entry point of various k8s services. We have
a dedicated dashboard on Logstash too, and we'd like to
have all access logs stored (if possible).

At the moment the traffic is low since not many services
run through the istio gateways, but the traffic will
surely ramp up in the future (for example when
the Lift Wing k8s infrastructure will takeover ORES).

This change adds a specific rule to avoid dropping traffic
from the proxy containers running in ingress gateways pods.

Change-Id: I0401a65b9d9b8cde0b3507fa705a29bfe2b9232a