varnish: add requestctl to X-analytics for static actions too
It happened repeatedly that some request pattern got blocked but the
corresponding requestctl rule wouldn't trigger or get logged.

The reason for that typically can be that one of the static rules is
matched before we get to requestctl generated rules.

In order to reduce the future confusion, add a requestctl marker
for each of these static rules, so that data in turnilo would be
easier to explain.

Bug: T342577
Change-Id: Ie0405655182492458960546289d48587af332f25