role::kafka::jumbo: apply thread settings
While doing perf tuning of the Kafka main clusters we realized
that num_io_threads and num_recovery_threads_per_data_dir were
not applied since the value was not passed between profile::kafka
and the confluent class (hence the hiera values got commented
to avoid new values to be picked up from all Kafka clusters
at once). Kafka jumbo is handling all the traffic really well,
despite the fact that it has only one thread (num.io.threads)
that processes all produce requests. Instead of using the old 12
value (that also refers to disks, and we have a hw raid set up
on jumbo), bump the value to 4 as we applied to other clusters
(in kafka main and logging the effects were visible also on
the codfw clusters were traffic is low).

Change-Id: I87d70c93e2532cb9f37febcbe844c4b472081523