releases: Change owner on /srv/patches after rsync
Every 10 minutes, we copy /srv/patches from the deploy* hosts to the
releases primary host. The releases puppet module requires /srv/patches
is owned by the jenkins user, but rsync maintains the owner (mwdeploy)
which is the same UID as debmonitor on the releases host.

This chown flag will stop puppet running chwon on the directory every
time it runs.

Bug: T342016
Change-Id: I54da9c9cb13516f86c223e48e8d331a2170e3002