Modify install and apt server config to support Juniper ZTP via HTTP
These changes are intended to adjust our install server DHCP config
to support using HTTP as the transport for Juniper ZTP.  This is
accomplished by adding some vendor-specific DHCP options for
Juniper to define the protocol used.  Another new var is used to
supply the JunOS image name that should be fetched.

For now I have statically set the values for the JunOS image, in
another iteration we should move this to the device-specific
config snippet generated from spicerack, and base the filename
on the Netbox platform parameter.  As we only need support for
QFX5120 devices short-term it is ok to hard-code as a first
version.

We will serve the files from our apt servers rather than the install
servers as previous.  To support this the 'tftp-server-name' supplied
by DHCP actually needs to provide the apt server IP.  Despite the
name of this option Juniper use it to provide the file server IP
irregardless of the actual protocol used.

The ztp-juniper.sh file fetched during the process also needs to be
created on the apt server now not the install server.

The apt server IP is retrieved by doing a DNS A lookup for
apt.discovery.wmnet (explictly A as we don't run v6 on the mgmt
network).

Bug: T336485
Change-Id: I4a5a4926e0b0c993ac46974203be66b51146818d