cumin: fix installer configuration
Why:
On bookworm the SSH with the installer key in the Debian installer is
pretty slow. It turns out this is caused because ssh tries first with
IPv6, then gets a "No route to host" and then retries with IPv4. This
can cause it to reach the timeout of 10 seconds currently used by
Spicerack to check the uptime of the host preventing it to detect the
host being in the Debian installer.

What:
Force the connection to be IPv4 only to prevent the unnecessary
slowness.

Hosts: auto
Bug: T342345
Change-Id: Ie4643b564886b4c4169f59a2d7585c8fe86540a2