idp_test: add datahub_staging as a OIDC service
We use a separate url for datahub-staging adding it as a registered service
and removing the datahub entry as we shall not be having the prod datahub
service on idp_test. This shall only be availed in prod after tests.
Corresponding secrets to be added when this is ready to merge.

Bug: T305874
Hosts: O:idp_test
Change-Id: I3a2281c6989c6e85566a7585add234805e166814