ferm::service: Fix handling of multiple ports
The delimiter for multiple ports was wrong, it needs to be a whitespace,
not a comma.

Also Ferm's syntax is a little peculiar: For a single port we can simply pass
the port, but if it's more than one we need to add an addiitonal layer of
brackets around it.

Hosts: C:profile::firewall
Bug: T336497
Change-Id: Idd9eb63b5840e9bf36a6b2fb7ebfa8ca97443486