Add IP pre-assignments for new lvs servers in Amsterdam
Documenting IPs reserved in Netbox for the new lvs servers being
installed in Amsterdam as part of the upcoming migration.

Have based server names of drmrs LVS, hardware is slightly different
(newer model), but it's not unlikely PCI placement will be the same.
We may of course need to update that when devices are online and
NICs are installed.

Bug: T343214
Change-Id: I2c3f3e15f9244cab1858cf1f62b5e2276252b4dc