Enable profile::auto_restarts::service for ipmiseld
Automatically restart it if underlying runtime library dependencies
have been upgraded. The restart is splayed across each day (on Mo-Fr).

Tested on bast2003 after the latest glibc update.

This was enabled in the past, but was reverted since some systems
failed to properly restart. It was unclear whether this was caused
by outdated firmware or a bug in ipmiseld or a combination of both.

So re-enable it on Bookworm only for now. These days we have
cookbook-driven firmware updates and thus generally more recent
deployed firmware and Bookworm also has a more recent ipmiseld version.

Bug: T135991
Hosts: bast2003.wikimedia.org
Change-Id: Ia1d5f42f67ef82c0db26ed54bd9cec86a995926d