bird::anycast_hc: temporarily remove validate_cmd
The initial Puppetization fails because validate_cmd can't find the
correct file. We should fix this in the Debian package so that we copy
the file to the right path where validate_cmd can temporarily find it,
pass the check, and then create the configuration file. For now, let's
disable it to check if it works as expected.

Change-Id: I7cb24516fab8636e4c140fadf7ab596a706032c4