Extend the firewall::service shim with checks for legacy syntax
There are a few cases of legacy syntax used in ferm::service definitions which
need to be updated before they can be used with nftables, catch these with an
explicit fail() so that they are visible in PCC when moving a role:

1. An srange or drange passed in Ferm-specific syntax like
srange => "@resolve((${hosts_allow_ferm}))"

2. A port passed in legacy port range syntax, e.g.
port   => '6800:7300'

3. A port passed in legacy Ferm syntax, e.g.
port   =>  '(4369 5671 5672 25672)'

4. Disabling connection tracking isn't implemented yet in the nftables config,
until then fail if used by a service.

Also add some typing to the shim where applicable and align the type for
the description.

Hosts: C:profile::firewall
Bug: T336497
Change-Id: I6f86aab061f251aba58f80dadb3e8b00733f7873