Preseed debian-installer not to prompt for additional firmware
We have previously attempted to preseed the hw-detect/load_firmware
debconf question to false, in order to prevent d-i from asking
interactively whether a user wishes to supply firmware.

However, the previous attempt didn't work because the question is asked
before the preseed file is retrieved. This is presumably in case you
need firmware to activate the network, prior to getting the preseed
file.

Adding the option to the kernel command line should get around this
problem.

Bug: T308106
Change-Id: I23784328e224f50218a6710ac150c67dab456376