P:bird::anycast: require anycast setup on the bird class
During the reimaging of durum hosts to bookworm (and also back to
bullseye, observed in both), we had a string of Puppet failures where it
couldn't find the `bird' user and thus resulted in errors like:

    Could not set 'directory' on ensure: Could not find user bird

This seems to happen because the anycast class is set up before the bird
one; the bird user is created when the bird2 package installed and that
happens in the bird module. Thus we should wait for that to be set up
before we start deploying the other anycast configuration bits.

Hosts: doh1001.wikimedia.org, dns1004.wikimedia.org
Change-Id: Iaefea9cee63b91a845858e604d601d291d32068a