P:bird::anycast: use systemd::sysuser for creating the bird user
In 574783904a1 and some other previous commits, we have been trying to
address the Puppet ordering for the bird::anycast profile to resolve
initial Puppetization of the anycast hosts. While we thought we had seen
the last of such issues, during the recent reimaging of the durum hosts
to bookworm, we observed that the bird user was not created and
therefore setting up anycast-healtchecker and bird's configuration files
was failing. Given that bird depends on the anycast-hc service, this was
leading to a series of failures preventing the host from being reimaged.

Fixing the ordering in the profile itself is a bit tricky. If we set up
anycast-hc before bird, the bird user does not exist as the bird2
package creates that in the postint. If we set up bird before
anycast-hc, that doesn't work as the sytemd bindings for bird prevent it
from starting (it After/BindsTo on the anycast-hc service).

The better fix for this then might be to create the user manually,
helping alleviate these issues, and then creating a dependency on that
user for both the anycast-hc and the bird stuff. There might be
additional dependencies between the anycast-hc and bird classes, but
this is a good start.

Change-Id: I4794f81389a76254e94d7399f66dd3b174e6e685