Use a routable email address for sending kerberos details
We have had reports of users receiving their kerberos details emails in
their spam folders. One of the possible causes of this is that we have
been sending the email with the From: header set to
root@krb1001.eqiad.wmnet which is non-routable.

This commit changes the script so that it sets the From: header to be
noc@wikimedia.org - It is hoped that this will improve the deliverability
of these messages.

In addition, we also remove an unused python module.

Bug: T318155
Change-Id: Iba64034122fdcebe976d64edf0086bf3a84e995e