Use the libmysql-java component on bullseye as well
We have previously been attempting to use the libmariadb-java connector
on bullseye, instead of the forward-ported libmysql-java from stretch.

It seems to have worked for sqoop, but it isn't working for hive so now
we need to start using this connector on bullseye as well.

I will follow this up with a change that will create a separate
bigop::mariadb_jdbc class, so that we can use this where it is known to
work (i.e. sqoop) and use MySQL where we still need to.

Bug: T329363
Change-Id: Ia141e715a29774a4a833c12a1748bda3cb5cb22c