bird: recursively manage the log file
The validate_cmd runs as root which creates a log file in the log folder
owned by root.  this prevents bird from starting as it tries to open the
same file with the bird user.  this Cr updates things we we ensure all
files are owned by bird before starting the service

Hosts: auto
Change-Id: I1c45bd8051b522025cb8ed8089c4348664868c24