admin: Create analytics-wmde system user and airflow admin group
Our standard practice for providing a team with an Airflow instance
requires the following:

* A system user and system group for the Airflow instance to use. This
  system user will be granted a kerberos keytab and will be able to
  submit jobs to the YARN cluster.

* A group of user accounts that are designated administrators of this
  instance. These users will have the rights to manage the Airflow
  services and to deploy the Airflow DAGs belonging to this team.

In the case of this instance for WMDE we have decided that the system
user and group will be named `analytics-wmde`.

The Airflow administrators for this instance will be named:
`airflow-wmde-admins`.

Approval for membership of the airflow-wmde-admins group will be the
responsbility of Kara Payne, the engineering manager for the WMDE group.

This change also restricts the journalctl rights for the various
airflow admin groups across all the airflow instances/groups.
The journalctl rights are now restricetd to only airflow-* services.

Bug: T340648
Bug: T345726
Change-Id: I18d4d6fb7478b8f318e59a54fd1ad3697a574e03