mw-debug-repl: Add --verbose
Why:
shell.php can produce verbose logging output (of MediaWiki logs)
using the --d 2 parameter, but it is not available in
mw-debug-repl context.

What:
Add --verbose to mw-debug-repl, which passes the debugging
parameter to shell.php.

Note: Completely untested, as I don't posses root
on deployment server to test this.

Bug: T344323
Change-Id: I8c3217d696d068b8c1109c69f24d1fab2c155911