Blacklist exfat
This is a filesystem we don't use on our servers (and won't ever need).

Blacklist it as defense in depth to reduce attack surface (e.g against container
images potentially using it), CVE-2023-4273 is an example of a Linux
vulnerability which would get prevented by this.

I've doublechecked with Cumin that none of our servers has the exfat kernel
module loaded.

Change-Id: Ie9017c9bb83b6e145fd3b884944e7e2c61595b9d