Grant analytics-admins rights to run some git cmds as analytics-deploy
This is in support of fixing the refinery deployment process, which has
been affected by the recent change to the behaviour of git with regard
to directory ownership.

The current `refinery-deploy-to-hdfs` script uses 'git describe' and the
procedure states that 'git log' should be used. Both of these are from a
scap deployment target directory, which cannot easily be added to the
git safe.directory global configuration.

This change will permit the members of analytics-admins to run the git
describe and git log commands as the analytics-deploy user.

A corresponding change to the script in the refinery repository will be
made in order to supply these sudo commands.

Bug: T334493
Hosts: P:analytics::refinery
Change-Id: I80f2dd521ab3d397ceb1816bc1d0fb431ae5e217