benthos: stop using strings as configuration
Given benthos allows us already to interpolate its programs with
environment variables that get injected, we don't really need to use
puppet templating on top of it - actually I think we should avoid it.

Add to this that this way we also make the puppet catalog leaner and
faster to compute, at just a small loss in terms of flexibility, which
I'd argue is a good thing in this case.

Change-Id: I4c81ff117ba53c6ed6fa1d6d4e339cd00b3d7ab9