confd: create run_dir via tmpfile
Having the timer depend on the file is not enough (I77f57d6aad) because
the timer might run at boot before puppet has a chance to run. Use
tmpfiles instead to make sure the directory exists at boot.

Another solution is to ConditionPathExists in the timer, however the
timer::job define doesn't support that.

Bug: T321678
Change-Id: If10b2a9cd284be2de1a19f2b749564f01a4534e7