esams: experimental frontend memory settings
As suggested by cdanis, we can speed up the trialing of new
frontend memory sizes by trying several values at once on
different nodes.  This sets the 8 nodes of each cluster in esams
to fe_mem_gb_reserved values from 80GB to 220GB in 20GB steps.
The nodes will need restarts after this patch (staggered to avoid
too much impact) to take effect.  The 220 value produces a
frontend memory size that approximates the legacy calculation,
while at the other end of the spectrum, the "80" nodes will
allocate ~140GB more frontend cache memory than the legacy value.

We want the "reserved" value to be as small as we can reasonable
get it without causing operational issues, but we probably need a
week or so of stability to feel confident the values isn't too
low.

Change-Id: I2e4972c3364d865a21451bf99fc39929169e076c