esams frontend memory: upload floor at least 120
cp3074 (set to 80) actually oomkilled varnish after several hours.
cp3075 (set to 100) seems to stabilize with ~11-12G of headroom at
the OS level, but that seems like too little, and it would
probably eventually oomkill during some future anomalous event.
So we'll try a floor of 120 on this cluster for now.

Change-Id: I735ed398678feb278127bb80560a9d7c57c1360f