esams: set frontend memory reservation to 128
Based on experimental evidence from using values ranging from
80-220 on both clusters, the most efficient and stable balance
seems like it should fall somewhere between the 120 and 140
settings.  128 is a nice round number in that range, so let's give
that a shot across a whole datacenter for extended validation of
stability (at least a week!) before spreading this to other sites.

For the latest-gen hosts with 512G total memory, this would set
the frontend malloc at 384G, which is a ~36% increase over the
legacy calculation of 283G on these hosts.  If the same value also
works with the (older-hardware, multi-backend) 384G sites, their
frontend malloc would raise by ~32% from 194G to 256G.

Change-Id: Ida5887006e4ab5a92fa6266d5f3f3a07aca6c5a4