idp_test: change datahub_staging profile_format
We have so far verified that authentication on the IDP side is okay and that we do
receive a signed id token. Datahub is however unable to locate OIDC attributes.
This is caused by how CAS returns the attributes. By default, the requested user profile is
rendered using a NESTED format where the authenticated principal and attributes are placed inside
id and attributes tags respectively in the final structure. Which is not expected by datahub.
We want ro try switch to FLAT where this option flattens principal attributes by one degree,
putting them at the same level as id. Other nested elements in the final payload are left untouched

Bug: T305874
Change-Id: Icb213e40d0f580fb5f690c74bc83c0d61d324ad3