Modify Juniper ZTP script used during initial provision
Corrects two issues:

1) Temp root password was not getting added to shell script due to
leaving the var declaration in the puppet class.

2) While testing with QFX5120-32C and a newer "out of the box" JunOS
version I hit an issue where it used virtual interface 'vme' to do
the initial DHCP.  This caused an issue as the previous script used
commands that referenced the interface name to get the DHCP
assigned IP and statically allocate it to em0.

Updated script gets the IP in a bit more complicated way, but it's
agnostic to the interface name and thus should work in all cases
we've seen.

Hosts: auto
Bug: T336485
Change-Id: I109e66cc9a7b93c8c9d31ce6831b3f2d218cbb05