Fix cache_upload timeouts in single-backend sites
When dc4f2bae ("hiera: Increase varnish max_connections to ats-be
on eqsin|ulsfo") was merged back in April of this year, the intent
was only to increase max_connections by a factor of 8 on the ATS
of both clusters (text+upload) at the new single-backend sites.

However, as an accidental consequence of the necessary copying and
pasting, the related timeout values from the text cluster were
copied to the upload clusters of these sites.  There have been
some intervening refactoring commits (and the conversion of esams
to single-backend) since which makes the history confusing, but
this commit should get it all back the way it should be.

Bug: T288106
Change-Id: Icb174700d403eb7d33a41bdab195d2c2c3c865d2