Move WMCS haproxy scrapes to WMCS prometheus instance
There are some firewall rules which block this kind of scraping from the
main ops instances. We will need to fix those later when moving the WMCS
instance to the main prometheus hardware, but for now move the scrape
rules to the 'cloud' instance which is a more correct place for those
anyways.

Bug: T345294
Change-Id: I17e093acad8380e7c4d769fbbe7fb7f837c7a843