Correct sysctl value for net.ipv4.tcp_min_snd_mss
We have this sysctl parameter set to mitigate an attack vector that
was revealed a few years back in CVE-2019-11479.

A user reported issues browsing Wikipedia from a system with an MTU
of 576 bytes, however.  I've also been able to reproduce this.

The issue is that 576 - 40 = 536, so if the intention is to support
MTUs as low as 576 this value should be 536.  Right now the lowest
MTU I can broswe Wikipedia from is 578.

We also should limit the lower MTU value we cache for destinations
to 40 bytes more than the minimum MSS we will use.  This prevents
us caching a MTU for a destination that is lower than the TCP
packets we will try to send.

Bug: T344829
Change-Id: I20269c0fcc669ced1527829334fb8bd253390bac