Add snapshot101[4-7] to the mediawiki-installation dsh group
Mediawiki is installed to these hosts in support of the XML Dumps
architecture, but at the moment are not in the mediawiki-installation
dsh group in puppet. This is generating alerts.

This commit adds these hosts to the dsh group with the intention of
clearing these alerts.

Hosts: O:dumps::generation::worker::testbed
Bug: T345907
Change-Id: Ie0f5b7b9473ad4312305ac347d75f73f120ce39c