profile::service_proxy::envoy: set use_ingress for thanos-swift
We are debugging an issue with connecting to Thanos Swift via
local mesh/proxy in k8s, and the leading theory seems to be
that envoy is not instructed correctly to set the SNI
when connecting to Swift on Thanos.

The use_ingress setting does exactly what we need, so since ML
is the only service using the thanos-swift listener it seems
safe to proceed. If this work a more permanent and stable fix
will be needed.

Bug: T339890
Change-Id: If549ac4d6779af5d18e5b33123220cbfc85efbff