k8s::apiserver: Use a separate systemd service for safe restarts
This reverts commit 45f585546c69204a0676afd6afe6c3540e084b3e as
ExecReload is not supposed to change the MAINPID of the systemd service
which makes this solution unreliable ad etcdctl will be terminated at
some point by systemd instead of actually waiting for the apiserver to
restart.

As an alternative approach a second systemd service
"kube-apiserver-safe-restart" is added (and notified by puppet instead
of "kube-apiserver") that acquires the etcd lock and restarts
kube-apiserver accordingly.

Bug: T329826
Hosts: auto
Change-Id: I9d591eac6a4e9328bccce008db5fb8dafef09f27