profile::thanos: add increase-based rec rules for Istio
The ML team is reviewing the metrics used to calculate
the SLO values, and we'd like to test increase().

The rationale is that we currently calculate the SLO
via avg_over_time(sum(rate)), meanwhile we'd like to
test sum(increase) to better calculate
all the events happened in a time range.

Bug: T327620
Change-Id: I98ae43e6637120d0d818d16be48ae6a3642fd338