external clouds: allow to get prefixes from RIPE
Why:
It would be useful to track public cloud IPs also for providers that
don't offer a standard way to get the list of prefixes for their IPs. In
those cases, as a fallback, we could use the set of publicly announced
prefixes for the given AS Number, for example from RIPE's APIs.
Although not perfect this could be a convenient method of getting
prefixes for any given AS Number.

What:
Add a new ExternalCloudVendorRIPE class to extract data from RIPE's
APIs, in particular from the stat.ripe.net website that offers
JSON output of announced prefixes for a given AS Number.
Configure it for Belcloud as a practical example.

Hosts: auto
Bug: T303534
Change-Id: I77a603fe66d1a19eeb2362f9ce0ffb1e017c235e