dynamicproxy: improve connection error pages
Show a more helpful message than the default 50x page generated by Nginx
when the backend webservice can't be reached. On Toolforge, that message
is customized so it's different depending on if it's a grid engine
webservice that can't be reached or if it's the main k8s ingress that's
unreachable.

Also all dynamicproxy error messages now include the hostname of the
instance that generated it to make it easier to find which component
generated which message.

Bug: T200616
Change-Id: I866351b72069f63e78f6a1ba9758c003cff443cb