fe_mem_gb_reserved:170 for all single-backend
This setting was tested for ~16 days in esams and seems to be
working well with a stable and expected amount of memory headroom.
eqsin and ulsfo are identical to esams in terms of hardware and
config, and so everything should work out the same at these sites
as well.  I suspect the same setting will also work at the legacy
multi-backend sites, but we'll save that for a later commit!

Change-Id: If925ccde6cde2495c338b1f5ef9d0d83abc1d9db