varnish: only listen on a single, local TCP port
We previously listened on 8x TCP ports for TIME_WAIT reasons.
We've since switched to using a single UDS for all runtime local
connections, so remove the excess listeners.  The remaining port
will be retained for debug/testing purposes, but is limited to
localhost to limit exposure and avoid misuse.

Bug: T333965
Change-Id: Icb0503114d802a888394652a66db4f7c5537b5ed