fe_mem_gb_reserved:170 for test hosts in other dcs
This tests the 170 reservation on 1 host each in the text+upload
clusters of the multi-backend DCs: eqiad, codfw, drmrs.  It's
expected to work out about the same as one the single-backend DCs,
based on the idea that the fixed memory headroom for everything
but the fe cache should be about the same, but best to verify with
a few hosts first!

The net change here is that varnish's malloc parameter raises from
194G to 207G (~+7%), on a host with a total 384G installed.

Change-Id: I68c363e5dd2b8d897754d48d48f4a989947c367e