Remove mention of an-test-client1001
As part of the Upgrade of the Hadoop test cluster to Bullseye, we recently onboarded a new test
client an-test-client1002 to replace the current an-test-client1001.
We now need to decommission an-test-client1001 and remove any mention of it from puppet.
This removes it from site.pp and from statistics_servers rsync allow group.

Bug: T329363
Bug: T343520
Change-Id: Ie694fa4e2785d4f7a2996df52cb525ba08e137c0