Add an nginx rule to block scripts from repositorygroup paths
We are aware of an issue with Archiva, whereby it allows scripts
appended to the URLs of repository groups to be executed in the visiting
clients browsers.

This commit adds a regex match to the nginx configuration, such that it
returns a 404 if an attempt is made to access a patch like this.

Bug: T318962
Change-Id: Ia8b73f7016635acb7aa96bb012d2d873ec381c8b