ferm: fix ferm-status on container bullseye instances
iptables-save can be found in /usr/sbin/iptables-save, but this script
looks for it in /sbin/iptables-save. This works fine on bookworm since
the official docker image has a merged /usr, but for bullseye /usr is
unmerged. As a solve just remove remove all hardcoded paths and rely on
the magic of PATH.

Bug: T344868
Hosts: basic
Change-Id: Ifeab98ef83b423cb678b77c04d64923453911ecc