openstack: eqiad1: pdns: refactor monitor checks
The host_check class is only in use by this profile. There is no
need for such decoupling, so merge the monitor checks into the main
profile.

Introduce a couple of new hiera keys, hopefully with enough of a
descriptive name.

And while at it, introduce modern `dnsquery::a()` instead of
`ipresolve()`.

Previous to this patch, the monitor checks were not working because
the new network setup.

Finally, to make it more useful as a monitoring tool, add support
to check more than one domain.

Bug: T346042
Hosts: re:cloudservices1.*
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: If965fa4fad5cad86c7a55999b1a60f875d3a4a81