varnish mem: unify global default
This spreads the now-well-tested 170G value to the remainder of
the multi-backend hosts in eqiad, codfw, and drmrs, while
consolidating all of the experimental hieradata settings back to
the class default.  This setting has been running for a while on
all single-backend hosts and 6 hosts (1 per cluster+site) at the
multi-backend sites as well.

Note also that the calculation was updated from the
legacy/deprecated fact $memorysize_mb to the modern equivalent
(does not change the final value).

Change-Id: Ic53b1363823db202f4d719b26538c545506239a5