Define a script in charge of checking the kafka broker in sync status
This script could then be used as a pre/post action of the
`sre.kafka.roll-restart-brokers` cookbook.

This way, we would get faster and/or safer rolling restarts.
Given that the current cookbook waits 900s between each restart,
either we move on to the next broker to restart as soon as the
current one is back in sync (modulo some artificial wait time),
or we prevent moving on to a next restart while the current broker
is still catching up.

Bug: T346741
Change-Id: Ia757b40da8ebf9e8d24cbed501196bb1fecc49cd
Hosts: kafka-jumbo1001.eqiad.wmnet