airflow-wmde: Remove statsd analytics-wmde user
The analytics-wmde user created for the purpose of running certain WMDE specific scripts on a
statsd host is no longer relevant. Instead we are creating a new system user analytics-wmde
that will be used to manage airflow and related jobs for the WMDE team.
To avoid conflicting users and permissions, we shall remove this initial user and
only have the system user.
WMDE team has given approval for this to take place, as the cronjob that was using the user
had alredy been disabled for a sigificant while during the spark3 upgrade.

Bug: T340648
Change-Id: I4cffb33301760dda410a17a60241cb14379cccf3