role::cache::text: set pass for ores.wikimedia.org
The ores domain has the same backend as ores-legacy,
and we are not ready to let Varnish to cache responses.
Several folks had issues with CORS headers, and we have been
struggling to make it work due to the current caching
strategy. The ML team would like to proceed in this way:
1) Set pass for ores too to solve the current inconsistency.
2) Work on setting the correct Vary/Cache headers in
   the backend code.
3) Replace pass with normal in the Varnish config.

Bug: T347344
Change-Id: I7e5db4e43319ea2c0933d2192f4467b3e9ff99e8