logspam-watch: Avoid termination via SIGPIPE
In function run_logspam, having `head` at the end of the pipeline was
causing the sort command to fail with SIGPIPE once `head` stopped
reading (but only if the output of output exceeded $listing_height).

This commit adds `cat >/dev/null` to the same subshell that `head` runs
in.  The cat will ensure that all output from `sort` is read to
completion.

Change-Id: If9843f45560676bdc443b7d95aa7ef9df7cbcd38