aptrepo: Add Bookworm HAProxy third party repos
This brings the Bookworm HAProxy packages to the same releases that
we're currently running on Bullseye.

haproxy.debian.net does not distribute 2.7; As only a few nodes are
running 2.7, Upgrading to 2.8 (the next LTS) seems the easiest way
forward.

This also removes the related Buster packages as we've since moved on to
Bullseye.

Bug: T342154
Change-Id: Ie10249983a6b5f2d98cc40b6734da103c836349c