mw-on-k8s: Remove wikidata exception
Why:
When migrating wikis to k8s based on group, we excluded wikidata because
of the amount of traffic compared to the rest of group1. Since we
changed to a global percentage based approach, this exclusion isn't
needed anymore.

What:
Remove m.wikidata.org and wikidata.org mw-on-k8s exclusion.

Bug: T290536
Change-Id: Iebdc15b19b652403dc4c51075bf51aff165b0802