cloudgw: refresh puppet namespace
Remove the openstack::$deployment namespace from the cloudgw roles and
profiles. We gain nothing from it.

Also, it gives the false indication that the cloudgw systems are per-deployment,
which is not the case.

This is similar to cloudlb, in the sense that this is generic infrastructure
for WMCS to use, independant of any openstack deployment or its shape.

No functional changes intended.

Bug: T347469
Hosts: re:cloudgw.*
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I7b26f972b0e043cfe25b96bf03ba478064aedf8b