install_server: replace ntp.$site with anycasted ntp.anycast.wmnet
Use the anycasted IP address for ntp.anycast.wmnet instead of the
per-site address, such as ntp.$site.wikimedia.org.

For a more detailed description of why we are doing this, please see the
task linked below. The TL;DR is that to reduce toil and manual updates
of the existing ntp.$site.wikimedia.org CNAME that points to a given DNS
host, we instead use the anycast address. By doing this, no updates to
the DNS records are necessary when doing maintenance on the DNS hosts.

Instead of having per-subnet overrides, we move this to common.cfg.

Hosts: O:apt_repo
Bug: T347054
Change-Id: I5f00d79c56c8cfa7883f1240ded4e9525ecbba3b