ci: manage cinder volume on Castor instance
integration-castor05 has an attached Cinder volume to store the CI
caches. I had them created with `prepare-cinder-volume` which adds an
fstab entry for it.

Today I have rebuild the instance and reattached the volume but it was
not mounted since the fstab entry was not on the new host. Add
`profile::labs::cindermount::srv` which detects attached volume and
mount them to `/srv`.

Since `/srv` is a Cinder volume we do not want to trigger the LVM logic
which manages extended disk space. The instance in horizon has:

  profile::ci::slave::labs::common::manage_srv: false

Pass it to the profile at the role level, this way we have all the logic
in Puppet and in Horizon we would solely apply the role.

Bug: T304080
Bug: T345924
Change-Id: I7926e97b11b8f9b26b5dd735bdff061c657c669d